ソースの整理

+++ src/TCTest.DomainService/DeleteUserDS.cs
@@ -11,6 +11,11 @@ public class DeleteUserDS(IUserRepository _userRepository) : IDeleteUserDS
 {
     public async Task ExecuteAsync(string userId)
     {
+        // 存在確認
+        if (await _userRepository.GetUserAsync(userId) == null)
+        {
+            throw new Exception("User not found.");
+        }
         await _userRepository.DeleteUserAsync(userId);
         await _userRepository.UnitOfWork.SaveEntitiesAsync();
     }
